progress on ray logic

+++ color.go
@@ -4,7 +4,7 @@ type Color struct {
 	R, G, B int
 }
 
-func WriteColor(r, g, b float32) Color {
+func WriteColor(r, g, b float64) Color {
 	ir := int(255.999 * r)
 	ig := int(255.999 * g)
 	ib := int(255.999 * b)
